pyproject.toml
[project]
name = "monday_report"
version = "0.1.0"
description = ""
authors = [
    {name = "Colin",email = "leocolin7002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "fastapi (>=0.118.3,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "types-pytz (>=2025.2.0.20250809,<2026.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.0"
mypy = "^1.18.2"

[tool.mypy]
ignore_missing_imports = true
strict_optional = true